requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.2
dadata==21.10.1
ecdsa==0.18.0
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.4.1
pydantic_core==2.10.1
pytest==8.0.1
pytest-asyncio==0.23.5
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.8
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
requests~=2.28.2
httpx==0.26.0
requests-mock==1.9.0
Jinja2==3.1.3
email_validator==2.1.1